Clank201 preview2 (#3)
Console should now congratulate you if you win. (untested, that's how bad I am at this game)
Cards now always fit in the screen, as long as they are not in a row with hidden cards.
Added some DeInits, so now pressing start should successfully close the game, and return to boot.3dsx
Renamed "data" folder to "gameData", so when building, the png files are not processed.